panta/pyproject.toml
[tool.poetry]
name = "lance"
authors = ["nashid, sijia"]
version = "0.0.0"
description = "Lance"
license = "Apache 2.0"
readme = "README.md"

packages = [
    { include = "lance" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
jinja2 = "^3.1.3"
beautifulsoup4 = "^4.12.3"
poetry-dynamic-versioning = "^1.3.0"

# LLM Dependencies
litellm = { git = "https://github.com/mrT23/litellm.git" }
tiktoken = "^0.7.0"
openai = "^1.32.0"
boto3 = "^1.34.121"
google-cloud-aiplatform = "^1.54.0"
numpy = "^1.26.0"
wandb = "^0.17.1"
dynaconf = "^3.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pyinstaller = "^6.6.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"
pytest-timeout = "^2.3.1"
fastapi = "^0.111.0"
networkx = "^2.6.3"
tree-sitter = "^0.20.1"
deepdiff = "^5.8.1"
pydot = "^1.4.1"
typer = "^0.12.3"
loguru = "^0.6.0"
numpy = "^1.26.4"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
lance = "lance.main:main"

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry-dynamic-versioning.from-file]
source = "lance/version.txt"
